Add FDE score. FSCK after installation. Various improvements..

+++ system/engine/feradroid.sh
@@ -11,6 +11,7 @@ CUR=$($B cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_cur_freq);
 CORES=$($B grep -c 'processor' /proc/cpuinfo);
 GPU=$(dumpsys SurfaceFlinger | $B grep "GLES:" | sed -e "s=GLES: ==");
 ARCH=$($B grep -Eo "ro.product.cpu.abi(2)?=.+" /system/build.prop 2>/dev/null | $B grep -Eo "[^=]*$" | head -n1);
+SCORE=/system/engine/prop/score;
 if [ -e /sys/power/cpufreq_max_axi_freq ]; then
  AXI=$($B cat /sys/power/cpufreq_max_axi_freq);
 fi;
@@ -24,22 +25,23 @@ if [ -e /sys/fs/selinux/enforce ]; then
  $B echo "0" > /sys/fs/selinux/enforce;
 fi;
 setprop persist.added_boot_bgservices "$CORES";
-setprop ro.config.max_starting_bg "$((CORES +1))";
-$B sleep 81;
+setprop ro.config.max_starting_bg "$((CORES+1))";
+svc power stayon true;
+$B sleep 72;
 svc power stayon true;
 setprop ro.feralab.engine 1.1;
-if [ -d /data/media/0/Android ]; then
- LOG=/data/media/0/Android/FDE_log.txt;
- CONFIG=/data/media/0/Android/FDE_config.txt;
-else
+if [ -d /sdcard/Android/ ]; then
  LOG=/sdcard/Android/FDE_log.txt;
  CONFIG=/sdcard/Android/FDE_config.txt;
+else
+ LOG=/data/media/0/Android/FDE_log.txt;
+ CONFIG=/data/media/0/Android/FDE_config.txt;
 fi;
 $B rm -f $LOG;
 $B touch $LOG;
 $B chmod 666 $LOG;
 $B chmod 666 $CONFIG;
-$B echo "50" > /sys/devices/virtual/timed_output/vibrator/enable;
+$B echo "60" > /sys/devices/virtual/timed_output/vibrator/enable;
 msg -t "FDE v1.1 - firing up...";
 {
  $B echo "### FeraLab ###"
@@ -74,16 +76,19 @@ msg -t "FDE v1.1 - firing up...";
 } >> $LOG;
 service call activity 51 i32 1;
 $B sleep 1;
-$B echo ">> Tuning SElinux.." >> $LOG;
-supolicy --live "allow sdcardd unlabeled dir { append create execute write relabelfrom link unlink ioctl getattr setattr read rename lock mounton quotaon swapon rmdir audit_access remove_name add_name reparent execmod search open }";
-supolicy --live "allow sdcardd unlabeled file { append create write relabelfrom link unlink ioctl getattr setattr read rename lock mounton quotaon swapon audit_access open }";
-supolicy --live "allow unlabeled unlabeled filesystem associate";
-supolicy --live "allow mediaserver mediaserver_tmpfs:file { read write execute }";
+if [ -e /sys/fs/selinux/enforce ]; then
+ $B echo ">> Tuning SElinux.." >> $LOG;
+ supolicy --live "allow sdcardd unlabeled dir { append create execute write relabelfrom link unlink ioctl getattr setattr read rename lock mounton quotaon swapon rmdir audit_access remove_name add_name reparent execmod search open }";
+ supolicy --live "allow sdcardd unlabeled file { append create write relabelfrom link unlink ioctl getattr setattr read rename lock mounton quotaon swapon audit_access open }";
+ supolicy --live "allow unlabeled unlabeled filesystem associate";
+ supolicy --live "allow mediaserver mediaserver_tmpfs:file { read write execute }";
+ supolicy --live "allow audioserver audioserver_tmpfs:file { read write execute }";
+fi;
 $B echo ">> Mounting partitions RW.." >> $LOG;
 mount -o remount,rw /data;
 mount -o remount,rw /system;
 mount -t debugfs none /sys/kernel/debug;
-mount debugfs debugfs /sys/kernel/debug;
+mount -t debugfs debugfs /sys/kernel/debug;
 $B chmod 0755 /sys/kernel/debug;
 if [ -e /sbin/sysrw ]; then
  $B echo ">> Remapped partition layout detected." >> $LOG;
@@ -116,6 +121,11 @@ if [ -e $CONFIG ]; then
  $B rm -f /system/engine/assets/FDE_config.txt;
  $B cp $CONFIG /system/engine/assets/FDE_config.txt;
 fi;
+$B echo "0" > $SCORE;
+$B echo "$((CORES+CORES+1))" >> $SCORE;
+if [ -e /sys/fs/selinux/enforce ]; then
+ $B echo "5" >> $SCORE;
+fi;
 if [ -e /system/engine/gears/dummy.sh ]; then
  DUMMY=$($B cat /system/engine/assets/FDE_config.txt | $B grep -e 'dummy=1');
  if [ "dummy=1" = "$DUMMY" ]; then
@@ -127,13 +137,16 @@ sync;
 $B sleep 1;
 $B echo ">> Executing kernel configuration..." >> $LOG;
 sysctl -p;
+$B echo "1" >> $SCORE;
 if [ "$SDK" -le "18" ]; then
  $B echo ">> Mediaserver fix..." >> $LOG;
  $B killall -9 android.process.media;
  $B killall -9 mediaserver;
+ $B echo "2" >> $SCORE;
 fi;
 if [ -e /etc/fstab ]; then
  $B echo "FStab onboard.";
+ $B echo "1" >> $SCORE;
 else
  $B cp /fstab.* /etc/fstab;
 fi;
@@ -143,6 +156,7 @@ $B echo ">> FileSystem trim..." >> $LOG;
 $B fstrim -v /system | $B tee -a $LOG;
 $B fstrim -v /data | $B tee -a $LOG;
 $B fstrim -v /cache | $B tee -a $LOG;
+$B echo "3" >> $SCORE;
 sync;
 if [ -e /system/engine/prop/firstboot ]; then
  mount -o remount,rw /system;
@@ -168,20 +182,30 @@ fi;
 setprop ro.secure 1;
 setprop ro.adb.secure 1;
 setprop security.perf_harden 1;
-setprop ro.debuggable 0;
+$B echo "3" >> $SCORE;
 $B echo ">> Tweaking multitasking.." >> $LOG;
 setprop ro.sys.fw.bg_apps_limit "$BG";
 service call activity 51 i32 "$BG";
+$B echo "$BG" >> $SCORE;
 svc power stayon false;
 $B echo "96" > /sys/devices/virtual/timed_output/vibrator/enable;
 $B sleep 0.3;
 $B echo "96" > /sys/devices/virtual/timed_output/vibrator/enable;
 $B sleep 0.3;
 $B echo "96" > /sys/devices/virtual/timed_output/vibrator/enable;
-msg -t "FDE status - OK.";
 $B killall -9 com.google.android.gms.persistent;
+$B echo "1" >> $SCORE;
+$B sleep 1;
+msg -t "FDE status - OK.";
 $B echo ">> FDE status - OK" >> $LOG;
 $B echo "  " >> $LOG;
+SCR=$($B awk '{ sum += $1 } END { print sum }' $SCORE);
+$B sleep 1;
+$B echo "*** Your FDE score is - $SCR ***" >> $LOG;
+msg -t "Your FDE score is - $SCR";
+$B echo "  " >> $LOG;
+$B sleep 1;
+msg -t "FDE LOG >> $LOG";
 if [ -e /engine.sh ]; then
  $B run-parts /system/etc/init.d;
 fi;
